repeated functions are abstracted out from common
